requirements.txt
# General Model
django >3.2,<3.3
python-decouple
pillow
djangorestframework

# Deplyment Module
gunicorn

# Date Time Module
jdatetime
jalali_core

# third party modules
drf-yasg
djangorestframework-simplejwt
django-filter
markdown


# Email Third Party Modules
django-mail-templated
django-cors-headers

# Refactoring
flake8
black

pytest
pytest-django

# Faker Data
Faker

# Request Modules
requests


# Background Process and Cache

celery
redis
django-redis